Reorganise JSON tests
